ISSUE #558: restructure gear namespaces
