+++ src/Domain/Strava/Gear/CustomGear/CustomGear.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace App\Domain\Strava\Gear\CustomGear;
 
-use App\Domain\Strava\Gear\ImportedGear;
+use App\Domain\Strava\Gear\ImportedGear\ImportedGear;
 use Doctrine\ORM\Mapping as ORM;
 
 #[ORM\Entity]
