+++ src/Domain/Strava/Gear/ImportedGear/ImportedGear.php
@@ -1,7 +1,10 @@
 <?php
 
-namespace App\Domain\Strava\Gear;
+namespace App\Domain\Strava\Gear\ImportedGear;
 
+use App\Domain\Strava\Gear\Gear;
+use App\Domain\Strava\Gear\GearId;
+use App\Domain\Strava\Gear\GearType;
 use App\Infrastructure\ValueObject\Measurement\Length\Kilometer;
 use App\Infrastructure\ValueObject\Measurement\Length\Meter;
 use App\Infrastructure\ValueObject\Time\SerializableDateTime;
