+++ src/Domain/Strava/Gear/ImportedGear/ImportedGearRepository.php
@@ -1,6 +1,9 @@
 <?php
 
-namespace App\Domain\Strava\Gear;
+namespace App\Domain\Strava\Gear\ImportedGear;
+
+use App\Domain\Strava\Gear\GearId;
+use App\Domain\Strava\Gear\Gears;
 
 interface ImportedGearRepository
 {
