+++ tests/Domain/Strava/Gear/CustomGear/DbalCustomGearRepositoryTest.php
@@ -6,10 +6,10 @@ use App\Domain\Strava\Gear\CustomGear\CustomGearRepository;
 use App\Domain\Strava\Gear\CustomGear\DbalCustomGearRepository;
 use App\Domain\Strava\Gear\GearId;
 use App\Domain\Strava\Gear\Gears;
-use App\Domain\Strava\Gear\ImportedGearRepository;
+use App\Domain\Strava\Gear\ImportedGear\ImportedGearRepository;
 use App\Infrastructure\ValueObject\Measurement\Length\Meter;
 use App\Tests\ContainerTestCase;
-use App\Tests\Domain\Strava\Gear\ImportedGearBuilder;
+use App\Tests\Domain\Strava\Gear\ImportedGear\ImportedGearBuilder;
 use Spatie\Snapshots\MatchesSnapshots;
 
 class DbalCustomGearRepositoryTest extends ContainerTestCase
